Updated VS project file, fixed compiler warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@6076 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/guiengine/widgets/dynamic_ribbon_widget.cpp
@@ -73,8 +73,8 @@ int estimateIconAreaFor(const int rowCount, const int wantedIconWidth,
 {
     const int row_height = height / rowCount;
 
-    float icon_height = row_height;
-    float icon_width = int(row_height * iconAspectRatio);
+    float icon_height = (float)row_height;
+    float icon_width = row_height * iconAspectRatio;
     
     const int icons_per_row = std::min(int(width / icon_width), int(width / wantedIconWidth));
